Move types from event.ts to types.ts.
